Modularize autocomplete launch
